did most of the other cdf stuff, tester works, but need to have a parser for 2008 data
