+++ src/cumulative_filer.py
@@ -8,10 +8,10 @@ import string
 df_cum = pd.read_csv("/Users/yipho/anes/cumulative_anes/test/3x_test.csv")
 df_cum.rename(columns={'2024 variables used 2x previous': '2024'}, inplace=True)
 
-df24 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/pdf_data/2024_pcodebook.csv")
-df20 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/pdf_data/2020_pcodebook.csv")
-df16 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/pdf_data/2016_pcodebook.csv")
-df12 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/pdf_data/2012_pcodebook.csv")
+df24 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/interim/2024_pcodebook.csv", encoding="utf-8")
+df20 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/interim/2020_pcodebook.csv", encoding="utf-8")
+df16 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/interim/2016_pcodebook.csv", encoding="utf-8")
+df12 = pd.read_csv("/Users/yipho/anes/cumulative_anes/data/interim/2012_pcodebook.csv", encoding="utf-8")
 
 def range_helper(var_string):
     var_pattern = re.compile(r"^(V\d+)([a-zA-Z])\-([a-zA-Z])$")
@@ -28,8 +28,6 @@ def range_helper(var_string):
         return [f"{base_var}{ch}" for ch in string.ascii_lowercase[start_idx:end_idx + 1]]
     else:
         return [var_str]
-
-
 
 rows = []
 for index, row in df_cum.iterrows():
@@ -89,11 +87,11 @@ for index, row in df_cum.iterrows():
             q_1 += row24['question'].iloc[0] + '\n'
 
             if pd.notna(row24['valid_labels'].iloc[0]):
-                Valid1 += row24['valid_labels'].iloc[0] + '\n'
+                Valid1 += str(row24['valid_labels'].iloc[0]) + '\n'
             else:
                 Valid1 += v24 + '\n'
             if pd.notna(row24['missing_labels'].iloc[0]):
-                Missing1 += row24['missing_labels'].iloc[0] + '\n'
+                Missing1 += str(row24['missing_labels'].iloc[0]) + '\n'
             else:
                 Missing1 += '\n'
         Sources += f"2024: {var_2024}\n"
@@ -117,11 +115,11 @@ for index, row in df_cum.iterrows():
             sum2 += row20['summary'].iloc[0] + '\n'
             q_2 += row20['question'].iloc[0] + '\n'
             if pd.notna(row20['valid_labels'].iloc[0]):
-                Valid2 += row20['valid_labels'].iloc[0] + '\n'
+                Valid2 += str(row20['valid_labels'].iloc[0]) + '\n'
             else:
                 Valid2 += '\n'
             if pd.notna(row20['missing_labels'].iloc[0]):
-                Missing2 += row20['missing_labels'].iloc[0] + '\n'
+                Missing2 += str(row20['missing_labels'].iloc[0]) + '\n'
             else:
                 Missing2 += '\n'
         Sources += f"2020: {var_2020}\n"
@@ -145,44 +143,79 @@ for index, row in df_cum.iterrows():
             if sum2:
                 sum3 += row16['summary'].iloc[0] + '\n'
                 q_3 += row16['question'].iloc[0] + '\n'
+
                 print(v16, row16['valid_labels'].iloc[0], type(row16['valid_labels'].iloc[0]))
-                Valid3 += row16['valid_labels'].iloc[0] + '\n'
-                Missing3 += row16['missing_labels'].iloc[0] + '\n'
+
+                if pd.notna(row16['valid_labels'].iloc[0]):
+                    Valid3 += str(row16['valid_labels'].iloc[0]) + '\n'
+                else:
+                    Valid3 += '\n'
+                if pd.notna(row16['missing_labels'].iloc[0]):
+                    Missing3 += str(row16['missing_labels'].iloc[0]) + '\n'
+                else:
+                    Missing3 += '\n'
             else:
                 sum2 += row16['summary'].iloc[0] + '\n'
                 q_2 += row16['question'].iloc[0] + '\n'
-                Valid2 += row16['valid_labels'].iloc[0] + '\n'
-                Missing2 += row16['missing_labels'].iloc[0] + '\n'
-        Sources += f"2016:{var_2016}\n"
+                if pd.notna(row16['valid_labels'].iloc[0]):
+                    Valid2 += str(row16['valid_labels'].iloc[0]) + '\n'
+                else:
+                    Valid2 += '\n'
+                if pd.notna(row16['missing_labels'].iloc[0]):
+                    Missing2 += str(row16['missing_labels'].iloc[0]) + '\n'
+                else:
+                    Missing2 += '\n'
+        Sources += f"2016: {var_2016}\n"
 
     if pd.notna(var_2012):
-        # if ";" in var_2012:
-        #     var_2012_list = [v12.strip() for v12 in var_2012.split(";")]
-        # else:
-        #     var_2012_list = [var_2012.strip()]
-        # for v12 in var_2012_list:
-        #     row12 = df12[df12['var_name'].str.strip() == v12]
-        #     if sum2:
-        #         sum3 += row12['summary'].iloc[0] + '\n'
-        #         q_3 += row12['question'].iloc[0] + '\n'
-        #         Valid3 += row12['valid_labels'].iloc[0] + '\n'
-        #         Missing3 += row12['missing_labels'].iloc[0] + '\n'
-        #     else:
-        #         sum2 += row12['summary'].iloc[0] + '\n'
-        #         q_2 += row12['question'].iloc[0] + '\n'
-        #         Valid2 += row12['valid_labels'].iloc[0] + '\n'
-        #         Missing2 += row12['missing_labels'].iloc[0] + '\n'
-        Sources += f"2012:{var_2012}\n"
+        raw_vars = [v.strip() for v in str(var_2012).split(";") if v.strip()]
+        print("Raw vars:", raw_vars)
+        expanded_vars = []
+        for v in raw_vars:
+            if '-' in v:
+                expanded = range_helper(v)
+                expanded_vars.extend(expanded)
+                print(expanded_vars)
+            else:
+                expanded_vars.append(v)
+
+        var_2012_list = expanded_vars
+
+        for v12 in var_2012_list:
+            row12 = df12[df12['var_name'].str.strip() == v12]
+            if sum2:
+                sum3 += row12['description'].iloc[0] + '\n'
+                q_3 += row12['description'].iloc[0] + '\n'
+                print(v16, row16['valid_labels'].iloc[0], type(row16['valid_labels'].iloc[0]))
+                if pd.notna(row12['valid_labels'].iloc[0]):
+                    Valid3 += str(row12['valid_labels'].iloc[0]) + '\n'
+                else:
+                    Valid3 += '\n'
+                if pd.notna(row12['missing_labels'].iloc[0]):
+                    Missing3 += str(row12['missing_labels'].iloc[0]) + '\n'
+                else:
+                    Missing3 += '\n'
+            else:
+                sum2 += row12['description'].iloc[0] + '\n'
+                q_2 += row12['description'].iloc[0] + '\n'
+                if pd.notna(row12['valid_labels'].iloc[0]):
+                    Valid2 += str(row12['valid_labels'].iloc[0]) + '\n'
+                else:
+                    Valid2 += '\n'
+                if pd.notna(row12['missing_labels'].iloc[0]):
+                    Missing2 += str(row12['missing_labels'].iloc[0]) + '\n'
+                else:
+                    Missing2 += '\n'
+        Sources += f"2012: {var_2012}\n"
 
     if pd.notna(var_2008):
-        Sources += f"2008:{var_2008}\n"
-
+        Sources += f"2008: {var_2008}\n"
     if pd.notna(var_1992):
-        Sources += f"1992:{var_1992}\n"
+        Sources += f"1992: {var_1992}\n"
     if pd.notna(var_1984):
-        Sources += f"1984:{var_1984}\n"
+        Sources += f"1984: {var_1984}\n"
     if pd.notna(var_1968):
-        Sources += f"1968:{var_1968}\n"
+        Sources += f"1968: {var_1968}\n"
 
     rows.append({
         'Variable': curr_var,
@@ -209,6 +242,6 @@ df_cdf = pd.DataFrame(rows)
 #                                'q_orig', 'q_1', 'q_2', 'q_3', 'Valid1', 'Valid2',
 #                                'Valid3', 'Missing1', 'Missing2', 'Missing3','Sources'])
 
-df_cdf.to_csv("/Users/yipho/anes/cumulative_anes/test/cdf_tester.csv", index=False)
+df_cdf.to_csv("/Users/yipho/anes/cumulative_anes/test/cdf_tester.csv", index=False, encoding = "utf-8-sig")
 
 # %%
