+++ src/pdf_extract.py
@@ -1,0 +1,25 @@
+"""
+Functions for extracting text from PDFs into .txt files.
+"""
+from pypdf import PdfReader
+import glob
+import os
+
+def pdf_extract(file_path):
+    reader = PdfReader(file_path)
+    all_text = ""
+    for page in reader.pages:
+        all_text += page.extract_text() + "\n"
+    base_name = os.path.splitext(os.path.basename(file_path))[0]
+    out_path = os.path.join(os.path.dirname(file_path), f"{base_name}.txt")
+
+    with open(out_path, "w", encoding="utf-8") as text_file:
+        text_file.write(all_text)
+
+    print(f"Saved: {out_path}")
+
+# Process all PDFs
+files = glob.glob("data/pdf_data/*.pdf", recursive=True)
+for file in files:
+    pdf_extract(file)
+
