+++ test/3x_test.csv
@@ -7,7 +7,7 @@ V242558,V202536,V162366,,,,,,Transgender people,,None of the 2016 and 2020 attit
 V242350,V202377,V162192,,,,,,Minimum wage should be increased/decreased,,
 V242317;V242318,V202329;V202330,V162146;V162147,,,,,,Favor/oppose requiring children vaccinated for public school,,
 V242354;V242355,V202381;V202382,V162161;V162162,,,,,,Do health benefits of vaccinations outweigh risks,,
-V242351;V242352,V202378;V202379,V162193;V192193a,,,,,,Govt should increase/decrease spending on health insurance,,
+V242351;V242352,V202378;V202379,V162193;V162193a,,,,,,Govt should increase/decrease spending on health insurance,,
 V242242,,V161204,aa_uni,,,,,Does R favor or oppose allowing universities to increase the number of black students by considering race...; How much favor/oppose,,
 V242523; V242524,V202491; V202492,V162320;V162321,,,,,,Do police treat whites or blacks better,,
 V242520; V242521,V202488; V202489,V162318;V162319,,,,,,Does federal govt treat whites or better better (typo on website),,
